fix: Pool stats display optimization
