+++ .gitignore
@@ -42,3 +42,5 @@ next-env.d.ts
 # IDE-specific files
 .idea
 .vscode
+
+.clinic
