+++ components/PoolStatsDisplay.tsx
@@ -1,7 +1,4 @@
-'use client';
-
-import { useEffect, useState } from 'react';
-
+import dynamic from 'next/dynamic';
 import Link from 'next/link';
 
 import { PoolStatsType } from '../lib/api';
@@ -14,6 +11,10 @@ import {
   getPercentageChangeColor,
   calculateAverageTimeToBlock,
 } from '../utils/helpers';
+
+const CountdownTimer = dynamic(() => import('./CountdownTimer'), {
+  ssr: false,
+});
 
 interface PoolStatsDisplayProps {
   stats: PoolStatsType;
@@ -186,34 +187,6 @@ export default function PoolStatsDisplay({
           </div>
         </div>
       </div>
-    </div>
-  );
-}
-
-function CountdownTimer({ initialSeconds }: { initialSeconds: number }) {
-  const [seconds, setSeconds] = useState(initialSeconds);
-
-  useEffect(() => {
-    const timer = setInterval(() => {
-      setSeconds((prevSeconds) => {
-        if (prevSeconds <= 1) {
-          clearInterval(timer);
-          window.location.reload();
-          return 0;
-        }
-        return prevSeconds - 1;
-      });
-    }, 1000);
-
-    return () => clearInterval(timer);
-  }, [initialSeconds]);
-
-  if (seconds === 0) {
-    return <div className="badge badge-primary">Updating Now</div>;
-  }
-  return (
-    <div className="badge badge-primary whitespace-nowrap">
-      Updating in {seconds}s
     </div>
   );
 }
